set up dev and inital ts conversion
