+++ .gitignore
@@ -1,2 +1,3 @@
 .env
+dist/
 node_modules/
