+++ package.json
@@ -5,8 +5,9 @@
   "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
-    "dev": "nodemon app.js"
+    "start": "node dist/app.js",
+    "start:dev": "nodemon dist/app.js",
+    "build": "tsc -p tsconfig.json"
   },
   "author": "Nwando Ifezue",
   "license": "ISC",
@@ -30,7 +31,16 @@
     "socket.io": "^4.7.2"
   },
   "devDependencies": {
+    "@types/body-parser": "^1.19.5",
+    "@types/express": "^5.0.0",
+    "@types/express-session": "^1.18.1",
+    "@types/node": "^22.13.1",
+    "@types/passport": "^1.0.17",
+    "@types/passport-facebook": "^3.0.3",
+    "@types/passport-google-oauth20": "^2.0.16",
+    "@types/passport-twitter": "^1.0.40",
     "mongoose-findorcreate": "^4.0.0",
-    "nodemon": "^3.0.2"
+    "nodemon": "^3.0.2",
+    "ts-node": "^10.9.2"
   }
 }
