--- public/css/htmlfile.css
@@ -1,63 +1,0 @@
-body {
-    background-color: rgb(85, 28, 99);
-    font-family: poppins;
-    color: white;
-}
-
-
-
-.navbar-brand{
-    font-family: Agbalumo;
-}
-
-.row{
-    padding: 10px 10px 60px 10px;
-}
-
-
-
-#login-buttons{
-    padding: 20px 10px 10px 0px;
-}
-
-
-#left-showcase{
-/* background-color: rgba(0, 0, 255, 0.363); */
-text-align: center;
-padding: 40px 25px 80px 10px;
-}
-
-#showcase-title h1{
-    font-size: 30px;
-    font-family:'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
-    font-weight: bold;
-}
-
-#showcase-text{
-    padding: 10px;
-}
-
-/* right showcase */
-#right-showcase{
-    background-color: rgba(165, 42, 42, 0.397);
-    padding: 20px;
-    text-align: center;
-    border-radius: 15px;
-    box-shadow: 0px -25px 20px -20px rgba(0, 0, 0, 0.45), 25px 0 20px -20px rgba(0, 0, 0, 0.45), 0px 25px 20px -20px rgba(0, 0, 0, 0.45), -25px 0 20px -20px rgba(0, 0, 0, 0.45);
-}
-
-#ceo-image img{
-    border-radius: 50%;
-    width: 40%;
-    padding-bottom: 10px;
-}
-
-footer{
-    text-align: center;
-    padding: 30px 0 20px 0;
-    background-color: rgb(64, 22, 75);
-}
-
-footer a {
-    text-decoration: none;
-}
