+++ src/views/discover.ejs
